src/backend/requirements.txt
asgiref==3.8.1
distlib==0.3.8
Django==4.1.13
djongo==1.3.6
dnspython==2.6.1
filelock==3.15.4
platformdirs==4.2.2
pymongo==3.12.3
sqlparse==0.2.4
tzdata==2024.1
virtualenv==20.26.3
djangorestframework
environs
django-cors-headers
pytz
Pillow
gunicorn